requirements.txt
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
discord.py==2.3.2
dnspython==2.4.0
exceptiongroup==1.1.3
flake8==7.0.0
flake8-print==5.0.0
flake8-quotes==3.4.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
logging==0.4.9.6
mccabe==0.7.0
Pillow==10.0.1
pycodestyle==2.11.1
pyflakes==3.2.0
pymongo==4.5.0
python-dotenv==1.0.0
reportlab==4.0.5
sniffio==1.3.0
typing_extensions==4.8.0